Don't even know, trying to do gragity field

+++ projects/project-1/shaders/particle-render.frag
@@ -1,8 +1,10 @@
-precision mediump float;
+precision highp float;
 
 varying float fLeft;  //a variable for each vertex?
 
 
 void main() {
+
   gl_FragColor = vec4(1.0, 1.0 - (fLeft * 0.8), 1.0 - (fLeft * 0.2), fLeft);
+
 }
